refactor: abstract feature flag webhook logic to lib
